[DEV-481] Use a different lock to protect t_continue()
The old LOCK_REPLIES in t_continue() would cause a deadlock in certain race
conditions in which t_reply() was also trying to aquire the replies lock
on a transaction.

Since the intention of the LOCK_REPLIES in t_continue() is to prevent
parallel calls to t_continue(), it is safer to use a different lock for
that purpose, called LOCK_CONTINUE. This way LOCK_REPLIES can be used as
intended without affecting t_continue().